convert bottom bar
